Add support for turning with gyro sensor

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/driving/DriveTrain.java
@@ -13,5 +13,5 @@ public abstract class DriveTrain extends HardwareComponent {
 
     public abstract void move(double millimeters, double degrees) throws InterruptedException;
 
-    public abstract void turn(double degrees) throws InterruptedException;
+    public abstract void turn(int degrees) throws InterruptedException;
 }
